requirements.txt
diffusers
datasets
transformers
accelerate
ftfy
pyarrow==19.0.0
matplotlib
torchvision
pytorch-fid